Extract filter processing logic to a seperate class
