+++ src/Filter/FilterProcessor.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * This file is part of the event-engine/php-postgres-document-store.
+ * (c) 2019-2021 prooph software GmbH <contact@prooph.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace EventEngine\DocumentStore\Postgres\Filter;
+
+use EventEngine\DocumentStore\Filter\Filter;
+
+interface FilterProcessor
+{
+    public function process(Filter $filter): FilterClause;
+}
